Silence Clang's `-Wformat` warnings
src/loader.c:408:79: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/loader.c:409:29: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/loader.c:409:36: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/loader.c:549:66: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/loader.c:549:71: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/loader.c:651:65: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/loader.c:832:75: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/loader.c:855:11: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/loader.c:855:19: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/loader.c:855:34: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/loader.c:855:49: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/loader.c:878:49: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/loader.c:879:49: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dis.c:63:81: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dis.c:73:81: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dis.c:140:42: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dis.c:140:63: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dis.c:377:13: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:137:40: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/dos.c:161:40: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/dos.c:210:63: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:285:41: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:285:61: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:286:31: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:309:67: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:476:44: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/dos.c:494:48: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/dos.c:807:63: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:1007:11: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:1007:23: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:1007:31: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:1007:39: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:1007:47: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:1007:55: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:1007:63: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:1711:44: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:1778:58: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:1806:49: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/dos.c:1818:56: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:1824:77: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:1848:65: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:1929:15: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:2284:69: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:2422:41: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/dos.c:2423:61: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/keyb.c:508:69: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/timer.c:150:17: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/video.c:86:54: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/video.c:86:63: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/video.c:92:73: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/video.c:92:82: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/video.c:214:49: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/video.c:264:55: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/video.c:264:61: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/video.c:359:39: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/video.c:367:43: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/video.c:411:42: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/video.c:515:65: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/video.c:626:53: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/video.c:626:64: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/video.c:627:11: warning: format specifies type 'int' but the argument has type 'unsigned int' [-Wformat]
src/video.c:817:73: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]
src/video.c:824:69: warning: format specifies type 'unsigned int' but the argument has type 'int' [-Wformat]

+++ src/keyb.c
@@ -505,7 +505,7 @@ uint8_t keyb_read_port(unsigned port)
 {
     if(queued_key == -1)
         kbhit();
-    debug(debug_int, "keyboard read_port: %02X (key=%04X)\n", port, queued_key);
+    debug(debug_int, "keyboard read_port: %02X (key=%04X)\n", port, (unsigned)queued_key);
     if(port == 0x60)
         return queued_key >> 8;
     else if(port == 0x61)
